backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-engineio==4.9.0
python-dotenv==1.0.1
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
pytest==8.0.0
pytest-flask==1.3.0
black==24.1.1
flake8==7.0.0
requests==2.32.3 
prometheus-client==0.20.0
flasgger==0.9.7.1
pydantic==2.7.4
aiohttp==3.9.1
cryptography==42.0.0
bcrypt==4.1.2
passlib==1.7.4
marshmallow==3.21.0
WeasyPrint==61.2

# Phase 9: Advanced Analytics and Machine Learning Dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0
yfinance==0.2.18
textblob==0.17.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
ta==0.10.2
nltk==3.8.1
transformers==4.33.2
torch==2.0.1
tensorflow==2.13.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2
optuna==3.3.0
mlflow==2.6.0
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
asyncio-mqtt==0.13.0
kafka-python==2.0.2
pyspark==3.4.1
dask==2023.8.1
vaex==4.17.0
streamlit==1.27.0
gradio==3.40.1
dash==2.14.0
bokeh==3.2.2
altair==5.1.2
holoviews==1.17.1
hvplot==0.9.2
panel==1.2.3
voila==0.4.0
jupyter==1.0.0
ipywidgets==8.0.7
nbformat==5.9.2
nbconvert==7.8.0
papermill==2.4.0
jupytext==1.15.1
nbsphinx==0.9.2
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
myst-parser==2.0.0
